feat(cli-zendesk): update logic to calculate the user fields "calculado"
Before we only counted in tickets that had a pending status. In this new logic, we calculate all tickets that are not solved, closed or deleted.
In cli-zendesk we odly only calculated these fields to MSR, not volunteers (our most valuable data to determine the match). Now we calculate these fields for MSR and Volunteers.