feat(mautic-zendesk): records registration_number again
In some mautic forms we dont ask for registration_number and assume that
the initial integration with zendesk got those info from the form_entry
and saved in the user.

Due to several problemas with this particular data, now we parse
registration_number from the form_entriy and save it in case the form
didnt have any. If the form has that field we just save that value, if
it doesnt we use the registration_number from the form_entry.